refactor(config): split resolver command owners
